+++ example/ios/Runner/AppDelegate.m
@@ -1,5 +1,6 @@
 #include "AppDelegate.h"
 #include "GeneratedPluginRegistrant.h"
+#include <TSBackgroundFetch/TSBackgroundFetch.h>
 
 @implementation AppDelegate
 
@@ -7,6 +8,9 @@
     didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
   [GeneratedPluginRegistrant registerWithRegistry:self];
   // Override point for customization after application launch.
+  TSBackgroundFetch *fetch = [TSBackgroundFetch sharedInstance];
+  [fetch registerBGProcessingTask:@"com.transistorsoft.customtask"];
+    
   return [super application:application didFinishLaunchingWithOptions:launchOptions];
 }