+++ CHANGELOG.md
@@ -1,3 +1,66 @@
+## 0.5.0 - 2020-01-03
+* [Added] [Android] New option `forceAlarmManager` for bypassing `JobScheduler` mechanism in favour of `AlarmManager` for more precise scheduling task execution.
+* [Changed] Migrate iOS deprecated "background-fetch" API to new [BGTaskScheduler](https://developer.apple.com/documentation/backgroundtasks/bgtaskscheduler?language=objc).  See new required steps in iOS Setup.
+* [Added] Added new `BackgroundFetch.scheduleTask` method for scheduling custom "onehot" and periodic tasks in addition to the default fetch-task.
+```dart
+BackgroundFetch.configure(BackgroundFetchConfig(
+  minimumFetchInterval: 15,
+  stopOnTerminate: false
+), (String taskId) {  // <-- [NEW] taskId provided to Callback
+  print("[BackgroundFetch] taskId: $taskId");
+  switch(taskId) {
+    case 'foo':
+      // Handle scheduleTask 'foo'
+      break;
+    default:
+      // Handle default fetch event.
+      break;
+  }
+  BackgroundFetch.finish(taskId);  // <-- [NEW] Provided taskId to #finish method.
+});
+
+// This event will end up in Callback provided to #configure above.
+BackgroundFetch.scheduleTask(TaskConfig(
+  taskId: 'foo',  //<-- required
+  delay: 60000,
+  periodic: false  
+));
+```
+
+## Breaking Changes
+* With the introduction of ability to execute custom tasks via `#scheduleTask`, all tasks are executed in the Callback provided to `#configure`.  As a result, this Callback is now provided an argument `String taskId`.  This `taskId` must now be provided to the `#finish` method, so that the SDK knows *which* task is being `#finish`ed.
+
+```dart
+BackgroundFetch.configure(BackgroundFetchConfig(
+  minimumFetchInterval: 15,
+  stopOnTerminate: false
+), (String taskId) {  // <-- [NEW] taskId provided to Callback
+  print("[BackgroundFetch] taskId: $taskId");
+  BackgroundFetch.finish(taskId);  // <-- [NEW] Provided taskId to #finish method.
+});
+```
+
+And with the Headless Task, as well:
+```dart
+/// This "Headless Task" is run when app is terminated.
+void backgroundFetchHeadlessTask(String taskId) async {  // <-- 1.  Headless task receives String taskId
+  print("[BackgroundFetch] Headless event received: $taskId");
+  
+  BackgroundFetch.finish(taskId);  // <-- 2.  #finish with taskId here as well.
+}
+
+void main() {
+  // Enable integration testing with the Flutter Driver extension.
+  // See https://flutter.io/testing/ for more info.
+  runApp(new MyApp());
+
+  // Register to receive BackgroundFetch events after app is terminated.
+  // Requires {stopOnTerminate: false, enableHeadless: true}
+  BackgroundFetch.registerHeadlessTask(backgroundFetchHeadlessTask);
+}
+``` 
+
+
 ## 0.4.0 - 2019-12-17
 * [Changed] Upgrade to new Flutter Plugin API "V2".  Requires flutter sdk version 1.12.  See [Upgrading pre 1.12 Android Projects](https://github.com/flutter/flutter/wiki/Upgrading-pre-1.12-Android-projects)